+++ br/pkg/lightning/importer/BUILD.bazel
@@ -14,7 +14,7 @@ go_library(
         "table_restore.go",
         "tidb.go",
     ],
-    importpath = "github.com/pingcap/tidb/br/pkg/lightning/restore",
+    importpath = "github.com/pingcap/tidb/br/pkg/lightning/importer",
     visibility = ["//visibility:public"],
     deps = [
         "//br/pkg/checksum",
