+++ br/pkg/lightning/importer/checksum_test.go
@@ -1,4 +1,4 @@
-package restore
+package importer
 
 import (
 	"context"
