+++ br/pkg/lightning/importer/mock/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "mock",
     srcs = ["mock.go"],
-    importpath = "github.com/pingcap/tidb/br/pkg/lightning/restore/mock",
+    importpath = "github.com/pingcap/tidb/br/pkg/lightning/importer/mock",
     visibility = ["//visibility:public"],
     deps = [
         "//br/pkg/lightning/mydump",
