+++ br/pkg/lightning/importer/opts/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
         "get_pre_info_opts.go",
         "precheck_opts.go",
     ],
-    importpath = "github.com/pingcap/tidb/br/pkg/lightning/restore/opts",
+    importpath = "github.com/pingcap/tidb/br/pkg/lightning/importer/opts",
     visibility = ["//visibility:public"],
     deps = ["//br/pkg/lightning/mydump"],
 )
