+++ br/tests/lightning_checkpoint_columns/run.sh
@@ -27,7 +27,7 @@ echo "INSERT INTO tbl (j, i) VALUES (3, 1),(4, 2);" > "$DBPATH/cp_tsr.tbl.sql"
 
 # Set minDeliverBytes to a small enough number to only write only 1 row each time
 # Set the failpoint to kill the lightning instance as soon as one row is written
-PKG="github.com/pingcap/tidb/br/pkg/lightning/restore"
+PKG="github.com/pingcap/tidb/br/pkg/lightning/importer"
 export GO_FAILPOINTS="$PKG/SlowDownWriteRows=sleep(1000);$PKG/FailAfterWriteRows=panic;$PKG/SetMinDeliverBytes=return(1)"
 # Check after 1 row is written in tidb backend, the finished progress is updated
 export GO_FAILPOINTS="${GO_FAILPOINTS};github.com/pingcap/tidb/br/pkg/lightning/PrintStatus=return()"
