+++ br/tests/lightning_checkpoint_timestamp/run.sh
@@ -19,7 +19,7 @@ set -eu
 run_sql 'DROP DATABASE IF EXISTS cpts'
 rm -f "$TEST_DIR"/cpts.pb*
 
-export GO_FAILPOINTS="github.com/pingcap/tidb/br/pkg/lightning/SetTaskID=return(1234567890);github.com/pingcap/tidb/br/pkg/lightning/restore/FailIfImportedChunk=return"
+export GO_FAILPOINTS="github.com/pingcap/tidb/br/pkg/lightning/SetTaskID=return(1234567890);github.com/pingcap/tidb/br/pkg/lightning/importer/FailIfImportedChunk=return"
 
 for i in $(seq 5); do
     echo "******** Importing Chunk Now (file step $i) ********"
