+++ br/tests/lightning_local_backend/run.sh
@@ -73,7 +73,7 @@ run_sql 'DROP DATABASE cpeng;'
 rm -f "/tmp/tidb_lightning_checkpoint_local_backend_test.pb"
 
 set +e
-export GO_FAILPOINTS='github.com/pingcap/tidb/br/pkg/lightning/restore/FailIfStatusBecomes=return(90);'
+export GO_FAILPOINTS='github.com/pingcap/tidb/br/pkg/lightning/importer/FailIfStatusBecomes=return(90);'
 for i in $(seq "$ENGINE_COUNT"); do
     echo "******** Importing Table Now (step $i/$ENGINE_COUNT) ********"
     run_lightning --backend local --enable-checkpoint=1 --log-file "$TEST_DIR/lightning-local.log" --config "tests/$TEST_NAME/config.toml"
@@ -106,7 +106,7 @@ for ckpt in mysql file; do
 
   # before chunk pos is updated, local files could handle lost
   set +e
-  export GO_FAILPOINTS="github.com/pingcap/tidb/br/pkg/lightning/restore/FailAfterWriteRows=return"
+  export GO_FAILPOINTS="github.com/pingcap/tidb/br/pkg/lightning/importer/FailAfterWriteRows=return"
   run_lightning --backend local --enable-checkpoint=1 --log-file "$TEST_DIR/lightning-local.log" --config "tests/$TEST_NAME/$ckpt.toml"
   set -e
   run_lightning_ctl --check-local-storage \
@@ -120,7 +120,7 @@ for ckpt in mysql file; do
   run_sql 'DROP DATABASE IF EXISTS tidb_lightning_checkpoint_local_backend_test'
   rm -f "/tmp/tidb_lightning_checkpoint_local_backend_test.pb"
   set +e
-  export GO_FAILPOINTS="github.com/pingcap/tidb/br/pkg/lightning/restore/LocalBackendSaveCheckpoint=return;github.com/pingcap/tidb/br/pkg/lightning/restore/FailIfImportedChunk=return"
+  export GO_FAILPOINTS="github.com/pingcap/tidb/br/pkg/lightning/importer/LocalBackendSaveCheckpoint=return;github.com/pingcap/tidb/br/pkg/lightning/importer/FailIfImportedChunk=return"
   run_lightning --backend local --enable-checkpoint=1 --log-file "$TEST_DIR/lightning-local.log" --config "tests/$TEST_NAME/$ckpt.toml"
   set -e
   run_lightning_ctl --check-local-storage \
@@ -133,7 +133,7 @@ for ckpt in mysql file; do
 
   # after index engine is imported, local file could handle lost
   set +e
-  export GO_FAILPOINTS="github.com/pingcap/tidb/br/pkg/lightning/restore/FailIfIndexEngineImported=return(1)"
+  export GO_FAILPOINTS="github.com/pingcap/tidb/br/pkg/lightning/importer/FailIfIndexEngineImported=return(1)"
   run_lightning --backend local --enable-checkpoint=1 --log-file "$TEST_DIR/lightning-local.log" --config "tests/$TEST_NAME/$ckpt.toml"
   set -e
   run_lightning_ctl --check-local-storage \
